Fix IPEX support and add XPU device to device_utils
