+++ sdxl_train_control_net_lllite.py
@@ -12,11 +12,9 @@ from types import SimpleNamespace
 import toml
 
 from tqdm import tqdm
+
 import torch
-
-from library.device_utils import clean_memory
-from library.ipex_interop import init_ipex
-
+from library.device_utils import init_ipex, clean_memory
 init_ipex()
 
 from torch.nn.parallel import DistributedDataParallel as DDP
