+++ train_textual_inversion_XTI.py
@@ -6,11 +6,9 @@ import toml
 from multiprocessing import Value
 
 from tqdm import tqdm
+
 import torch
-
-from library.device_utils import clean_memory
-from library.ipex_interop import init_ipex
-
+from library.device_utils import init_ipex, clean_memory
 init_ipex()
 
 from accelerate.utils import set_seed
